edit generatePublicKey in rmb-client
