+++ packages/rmb_direct_client/lib/client.ts
@@ -397,3 +397,4 @@ class Client {
   }
 }
 export { Client };
+export * from "./util";
